+++ src/elixir/String/index.ts
@@ -1,4 +1,5 @@
 export { downcase } from './downcase.ts'
 export { length } from './length.ts'
+export { replace } from './replace.ts'
 export { reverse } from './reverse.ts'
 export { upcase } from './upcase.ts'